backend/requirements.txt
# common
dotenv
boto3
pandas
fastapi

#growjo scraper
selenium
bs4
webdriver_manager

#mistral ocr
mistralai>=0.0.11
google-generativeai
pinecone[grpc]
langchain
langgraph

nltk
sentence_transformers

playwright

pyarrow<19.0.0
snowflake-connector-python

supabase
bcrypt
python-multipart

tavily-python
mcp[cli]
snowflake-snowpark-python
sqlparse

mcp
mcp-agent
openai
openai-tools
openai-agents
openai-agents-mcp

tensorflow
keras
tf-keras

matplotlib
plotly
